+++ bookprintapi/__init__.py
@@ -1,13 +1,21 @@
 """BookPrintAPI Python SDK"""
 
 from .client import Client
-from .exceptions import ApiError, FieldError, ValidationError
+from .exceptions import (
+    ApiError,
+    FieldError,
+    HelperErrorCodes,
+    HelperStage,
+    SweetbookHelperError,
+    ValidationError,
+)
+from .helpers import BookBuildResult, HelpersClient, PdfOrderBuildResult
 from .response import ResponseParser
 from .errorcodes import ErrorCodes, ConstraintTypes
 from .order_status import OrderStatus, ORDER_STATUS_CODE, ORDER_STATUS_FROM_CODE
 from .webhook import verify_signature
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 __all__ = [
     "Client",
     "ApiError",
@@ -20,4 +28,11 @@ __all__ = [
     "ORDER_STATUS_CODE",
     "ORDER_STATUS_FROM_CODE",
     "verify_signature",
+    # Helpers
+    "BookBuildResult",
+    "PdfOrderBuildResult",
+    "HelpersClient",
+    "SweetbookHelperError",
+    "HelperStage",
+    "HelperErrorCodes",
 ]